% tinyrox says don't edit this manually, but it can't stop you!
\name{active_media_reference_key}
\alias{active_media_reference_key}
\alias{active_media_reference_key<-}
\title{Active media reference key of a Clip}
\usage{
active_media_reference_key(x)

active_media_reference_key(x) <- value
}
\arguments{
\item{x}{A \code{Clip}.}

\item{value}{The key to make active (must exist).}
}
\value{
The active key string; the setter returns \code{x}.
}
\description{
Active media reference key of a Clip
}
