% tinyrox says don't edit this manually, but it can't stop you!
\name{children}
\alias{children}
\title{Children of a composition}
\usage{
children(x)
}
\arguments{
\item{x}{A \code{Composition} (Track, Stack) or \code{SerializableCollection}.}
}
\value{
A list of child OTIO objects, in order.
}
\description{
Children of a composition
}
