% tinyrox says don't edit this manually, but it can't stop you!
\name{color}
\alias{color}
\alias{color<-}
\title{Marker color}
\usage{
color(x)

color(x) <- value
}
\arguments{
\item{x}{A \code{Marker}.}

\item{value}{A color string (e.g. \code{"RED"}, \code{"GREEN"}).}
}
\value{
The color string; the setter returns \code{x}.
}
\description{
Marker color
}
