% tinyrox says don't edit this manually, but it can't stop you!
\name{default_media_key}
\alias{default_media_key}
\title{The default media reference key}
\usage{
default_media_key()
}
\value{
The string OTIO uses for a clip's default media (\code{"DEFAULT_MEDIA"}).
}
\description{
The default media reference key
}
