% tinyrox says don't edit this manually, but it can't stop you!
\name{Effect}
\alias{Effect}
\title{Effect}
\usage{
Effect(name = "", effect_name = "", enabled = TRUE, metadata = list())
}
\arguments{
\item{name}{Object name.}

\item{effect_name}{Effect identifier string.}

\item{enabled}{Whether the effect is active.}

\item{metadata}{Named list of metadata.}
}
\value{
An \code{Effect}.
}
\description{
Effect
}
