% tinyrox says don't edit this manually, but it can't stop you!
\name{end_time_exclusive}
\alias{end_time_exclusive}
\alias{end_time_inclusive}
\title{Exclusive and inclusive end times of a TimeRange}
\usage{
end_time_exclusive(tr)

end_time_inclusive(tr)
}
\arguments{
\item{tr}{A \code{TimeRange}.}
}
\value{
A \code{RationalTime}.
}
\description{
Exclusive and inclusive end times of a TimeRange
}
