% tinyrox says don't edit this manually, but it can't stop you!
\name{ExternalReference}
\alias{ExternalReference}
\title{ExternalReference}
\usage{
ExternalReference(target_url = "", available_range = NULL, metadata = list())
}
\arguments{
\item{target_url}{URL or path to the media.}

\item{available_range}{A \code{TimeRange} or \code{NULL}.}

\item{metadata}{Named list of metadata.}
}
\value{
An \code{ExternalReference}.
}
\description{
ExternalReference
}
