% tinyrox says don't edit this manually, but it can't stop you!
\name{find_clips}
\alias{find_clips}
\title{Find all clips within an object (recursively)}
\usage{
find_clips(x)
}
\arguments{
\item{x}{A \code{Composition}, \code{Timeline}, or \code{SerializableCollection}.}
}
\value{
A list of \code{Clip} objects.
}
\description{
Find all clips within an object (recursively)
}
