% tinyrox says don't edit this manually, but it can't stop you!
\name{FreezeFrame}
\alias{FreezeFrame}
\title{FreezeFrame}
\usage{
FreezeFrame(name = "", metadata = list())
}
\arguments{
\item{name}{Object name.}

\item{metadata}{Named list of metadata.}
}
\value{
A \code{FreezeFrame}.
}
\description{
FreezeFrame
}
\examples{
ff <- FreezeFrame("hold")
}
