% tinyrox says don't edit this manually, but it can't stop you!
\name{from_frames}
\alias{from_frames}
\title{Build a RationalTime from a frame number}
\usage{
from_frames(frame, rate)
}
\arguments{
\item{frame}{Frame number.}

\item{rate}{Frame rate.}
}
\value{
A \code{RationalTime}.
}
\description{
Build a RationalTime from a frame number
}
\examples{
from_frames(48, 24)
}
