% tinyrox says don't edit this manually, but it can't stop you!
\name{from_seconds}
\alias{from_seconds}
\title{Build a RationalTime from seconds}
\usage{
from_seconds(seconds, rate = 1)
}
\arguments{
\item{seconds}{Numeric seconds.}

\item{rate}{Target sample rate.}
}
\value{
A \code{RationalTime}.
}
\description{
Build a RationalTime from seconds
}
