% tinyrox says don't edit this manually, but it can't stop you!
\name{Gap}
\alias{Gap}
\title{Gap}
\usage{
Gap(duration, name = "", metadata = list())
}
\arguments{
\item{duration}{A \code{RationalTime} giving the gap length.}

\item{name}{Object name.}

\item{metadata}{Named list of metadata.}
}
\value{
A \code{Gap}.
}
\description{
Gap
}
