% tinyrox says don't edit this manually, but it can't stop you!
\name{has_child}
\alias{has_child}
\title{Does a composition directly contain a child?}
\usage{
has_child(x, child)
}
\arguments{
\item{x}{A \code{Composition}.}

\item{child}{A \code{Composable}.}
}
\value{
Logical.
}
\description{
Does a composition directly contain a child?
}
