% tinyrox says don't edit this manually, but it can't stop you!
\name{has_clips}
\alias{has_clips}
\title{Does a composition contain any clips (recursively)?}
\usage{
has_clips(x)
}
\arguments{
\item{x}{A \code{Composition}.}
}
\value{
Logical.
}
\description{
Does a composition contain any clips (recursively)?
}
