% tinyrox says don't edit this manually, but it can't stop you!
\name{insert_child}
\alias{insert_child}
\title{Insert a child at a position}
\usage{
insert_child(x, index, child)
}
\arguments{
\item{x}{A \code{Composition}.}

\item{index}{1-based position at which to insert.}

\item{child}{A \code{Composable}.}
}
\value{
\code{x}, invisibly.
}
\description{
Insert a child at a position
}
