% tinyrox says don't edit this manually, but it can't stop you!
\name{is_equivalent_to}
\alias{is_equivalent_to}
\title{Object-value equivalence (not pointer identity)}
\usage{
is_equivalent_to(x, other)
}
\arguments{
\item{x,other}{OTIO objects.}
}
\value{
Logical.
}
\description{
Object-value equivalence (not pointer identity)
}
