% tinyrox says don't edit this manually, but it can't stop you!
\name{is_unknown_schema}
\alias{is_unknown_schema}
\title{Was this object loaded from an unknown schema?}
\usage{
is_unknown_schema(x)
}
\arguments{
\item{x}{An OTIO object.}
}
\value{
Logical.
}
\description{
Was this object loaded from an unknown schema?
}
