% tinyrox says don't edit this manually, but it can't stop you!
\name{LinearTimeWarp}
\alias{LinearTimeWarp}
\title{LinearTimeWarp}
\usage{
LinearTimeWarp(name = "", effect_name = "", time_scalar = 1, metadata = list())
}
\arguments{
\item{name}{Object name.}

\item{effect_name}{Effect identifier string.}

\item{time_scalar}{Playback rate multiplier (2 = double speed).}

\item{metadata}{Named list of metadata.}
}
\value{
A \code{LinearTimeWarp}.
}
\description{
LinearTimeWarp
}
