% tinyrox says don't edit this manually, but it can't stop you!
\name{Marker}
\alias{Marker}
\title{Marker}
\usage{
Marker(name = "", marked_range = TimeRange(), color = "GREEN", comment = "",
       metadata = list())
}
\arguments{
\item{name}{Object name.}

\item{marked_range}{A \code{TimeRange}.}

\item{color}{Marker color string (e.g. \code{"GREEN"}, \code{"RED"}).}

\item{comment}{Free-text comment.}

\item{metadata}{Named list of metadata.}
}
\value{
A \code{Marker}.
}
\description{
Marker
}
