% tinyrox says don't edit this manually, but it can't stop you!
\name{media_reference}
\alias{media_reference}
\alias{media_reference<-}
\title{Active media reference of a Clip}
\usage{
media_reference(x)

media_reference(x) <- value
}
\arguments{
\item{x}{A \code{Clip}.}

\item{value}{A media reference object or \code{NULL}.}
}
\value{
The active media reference (or \code{NULL}); the setter returns \code{x}.
}
\description{
Active media reference of a Clip
}
