% tinyrox says don't edit this manually, but it can't stop you!
\name{MediaReference}
\alias{MediaReference}
\title{MediaReference}
\usage{
MediaReference(name = "", available_range = NULL, metadata = list())
}
\arguments{
\item{name}{Object name.}

\item{available_range}{A \code{TimeRange} or \code{NULL}.}

\item{metadata}{Named list of metadata.}
}
\value{
A \code{MediaReference}.
}
\description{
MediaReference
}
