% tinyrox says don't edit this manually, but it can't stop you!
\name{name_prefix}
\alias{name_prefix}
\alias{name_suffix}
\alias{start_frame}
\alias{frame_step}
\alias{frame_zero_padding}
\alias{end_frame}
\alias{number_of_images_in_sequence}
\title{Image sequence frame fields}
\usage{
name_prefix(x)

name_suffix(x)

start_frame(x)

frame_step(x)

frame_zero_padding(x)

end_frame(x)

number_of_images_in_sequence(x)
}
\arguments{
\item{x}{An \code{ImageSequenceReference}.}
}
\value{
The named field value.
}
\description{
Image sequence frame fields
}
