% tinyrox says don't edit this manually, but it can't stop you!
\name{overlapping}
\alias{overlapping}
\title{Does a Composable overlap its neighbours? (TRUE for transitions)}
\usage{
overlapping(x)
}
\arguments{
\item{x}{A \code{Composable}.}
}
\value{
Logical.
}
\description{
Does a Composable overlap its neighbours? (TRUE for transitions)
}
