% tinyrox says don't edit this manually, but it can't stop you!
\name{range_from_start_end_time}
\alias{range_from_start_end_time}
\title{Build a TimeRange from start and exclusive end times}
\usage{
range_from_start_end_time(start_time, end_time_exclusive)
}
\arguments{
\item{start_time,end_time_exclusive}{\code{RationalTime} objects.}
}
\value{
A \code{TimeRange}.
}
\description{
Build a TimeRange from start and exclusive end times
}
