% tinyrox says don't edit this manually, but it can't stop you!
\name{rescaled_to}
\alias{rescaled_to}
\title{Rescale a RationalTime to a new rate}
\usage{
rescaled_to(rt, new_rate)
}
\arguments{
\item{rt}{A \code{RationalTime}.}

\item{new_rate}{Target rate.}
}
\value{
A \code{RationalTime} at \code{new_rate}.
}
\description{
Rescale a RationalTime to a new rate
}
