% tinyrox says don't edit this manually, but it can't stop you!
\name{SerializableCollection}
\alias{SerializableCollection}
\title{SerializableCollection}
\usage{
SerializableCollection(name = "", children = list(), metadata = list())
}
\arguments{
\item{name}{Object name.}

\item{children}{A list of OTIO objects.}

\item{metadata}{Named list of metadata.}
}
\value{
A \code{SerializableCollection}.
}
\description{
SerializableCollection
}
