% tinyrox says don't edit this manually, but it can't stop you!
\name{target_url_base}
\alias{target_url_base}
\alias{target_url_base<-}
\title{Base URL of an ImageSequenceReference}
\usage{
target_url_base(x)

target_url_base(x) <- value
}
\arguments{
\item{x}{An \code{ImageSequenceReference}.}

\item{value}{A base URL string.}
}
\value{
The base URL; the setter returns \code{x}.
}
\description{
Base URL of an ImageSequenceReference
}
