% tinyrox says don't edit this manually, but it can't stop you!
\name{target_url_for_image_number}
\alias{target_url_for_image_number}
\title{Target URL for a given image number in a sequence}
\usage{
target_url_for_image_number(x, image_number)
}
\arguments{
\item{x}{An \code{ImageSequenceReference}.}

\item{image_number}{1-based image number within the sequence.}
}
\value{
A URL string.
}
\description{
Target URL for a given image number in a sequence
}
