% tinyrox says don't edit this manually, but it can't stop you!
\name{time_scalar}
\alias{time_scalar}
\alias{time_scalar<-}
\title{Time scalar of a LinearTimeWarp}
\usage{
time_scalar(x)

time_scalar(x) <- value
}
\arguments{
\item{x}{A \code{LinearTimeWarp}.}

\item{value}{Numeric playback rate multiplier.}
}
\value{
The time scalar; the setter returns \code{x}.
}
\description{
Time scalar of a LinearTimeWarp
}
