% tinyrox says don't edit this manually, but it can't stop you!
\name{TimeEffect}
\alias{TimeEffect}
\title{TimeEffect}
\usage{
TimeEffect(name = "", effect_name = "", metadata = list())
}
\arguments{
\item{name}{Object name.}

\item{effect_name}{Effect identifier string.}

\item{metadata}{Named list of metadata.}
}
\value{
A \code{TimeEffect}.
}
\description{
TimeEffect
}
\examples{
te <- TimeEffect("retime", effect_name = "Retime")
}
