% tinyrox says don't edit this manually, but it can't stop you!
\name{to_json_string}
\alias{to_json_string}
\title{Serialize an OTIO object to a JSON string}
\usage{
to_json_string(x, indent = 4L)
}
\arguments{
\item{x}{An OTIO object.}

\item{indent}{Indentation width (spaces).}
}
\value{
A JSON string in OTIO's vocabulary.
}
\description{
Serialize an OTIO object to a JSON string
}
