% tinyrox says don't edit this manually, but it can't stop you!
\name{to_time_string}
\alias{to_time_string}
\alias{from_time_string}
\title{Time-string conversions (HH:MM:SS.s)}
\usage{
to_time_string(rt)

from_time_string(time_string, rate)
}
\arguments{
\item{rt}{A \code{RationalTime}.}

\item{time_string}{A \code{HH:MM:SS.s} string.}

\item{rate}{Sample rate.}
}
\value{
A time string.
}
\description{
Time-string conversions (HH:MM:SS.s)
}
