% tinyrox says don't edit this manually, but it can't stop you!
\name{clear_children}
\alias{clear_children}
\title{Remove all children of a composition}
\usage{
clear_children(x)
}
\arguments{
\item{x}{A \code{Composition} or \code{SerializableCollection}.}
}
\value{
\code{x}, invisibly.
}
\description{
Remove all children of a composition
}
