% tinyrox says don't edit this manually, but it can't stop you!
\name{flatten_stack}
\alias{flatten_stack}
\title{Flatten a stack (or list of tracks) into a single track}
\usage{
flatten_stack(x)
}
\arguments{
\item{x}{A \code{Stack}, or a list of \code{Track} objects.}
}
\value{
A new \code{Track} with the layers composited top-down.
}
\description{
Flatten a stack (or list of tracks) into a single track
}
