% tinyrox says don't edit this manually, but it can't stop you!
\name{frame_for_time}
\alias{frame_for_time}
\title{Frame number covering a time in an image sequence}
\usage{
frame_for_time(x, time)
}
\arguments{
\item{x}{An \code{ImageSequenceReference}.}

\item{time}{A \code{RationalTime}.}
}
\value{
Integer frame number.
}
\description{
Frame number covering a time in an image sequence
}
