% tinyrox says don't edit this manually, but it can't stop you!
\name{index_of_child}
\alias{index_of_child}
\title{Position of a child within a composition}
\usage{
index_of_child(x, child)
}
\arguments{
\item{x}{A \code{Composition}.}

\item{child}{A \code{Composable}.}
}
\value{
The 1-based index, or \code{NA} if not a child.
}
\description{
Position of a child within a composition
}
