% tinyrox says don't edit this manually, but it can't stop you!
\name{is_parent_of}
\alias{is_parent_of}
\title{Is one object a parent (ancestor) of another?}
\usage{
is_parent_of(x, other)
}
\arguments{
\item{x}{A \code{Composition}.}

\item{other}{A \code{Composable}.}
}
\value{
Logical.
}
\description{
Is one object a parent (ancestor) of another?
}
