% tinyrox says don't edit this manually, but it can't stop you!
\name{parent}
\alias{parent}
\title{Parent composition of a Composable}
\usage{
parent(x)
}
\arguments{
\item{x}{A \code{Composable}.}
}
\value{
The parent \code{Composition}, or \code{NULL} if unparented.
}
\description{
Parent composition of a Composable
}
