% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{rotio-package}
\alias{rotio}
\alias{rotio-package}
\title{\packageTitle{rotio}}
\description{\packageDescription{rotio}}
\details{
Names follow OTIO exactly: \code{source_range}, \code{range_in_parent},
\code{append_child}, and so on. Tree children are addressed by position
and object reference, never by name.

}
\section{Package Content}{\packageIndices{rotio}}
\author{\packageAuthor{rotio}}
\section{Maintainer}{\packageMaintainer{rotio}}
\keyword{internal}
