% tinyrox says don't edit this manually, but it can't stop you!
\name{set_child}
\alias{set_child}
\title{Replace the child at a position}
\usage{
set_child(x, index, child)
}
\arguments{
\item{x}{A \code{Composition}.}

\item{index}{1-based position to replace.}

\item{child}{A \code{Composable}.}
}
\value{
\code{x}, invisibly.
}
\description{
Replace the child at a position
}
