% tinyrox says don't edit this manually, but it can't stop you!
\name{to_json_file}
\alias{to_json_file}
\title{Write an OTIO object to a JSON file}
\usage{
to_json_file(x, file_name, indent = 4L)
}
\arguments{
\item{x}{An OTIO object.}

\item{file_name}{Destination path.}

\item{indent}{Indentation width (spaces).}
}
\value{
\code{TRUE} on success (invisibly).
}
\description{
Write an OTIO object to a JSON file
}
