% tinyrox says don't edit this manually, but it can't stop you!
\name{visible_range}
\alias{visible_range}
\title{Visible range of an Item (including handles into transitions)}
\usage{
visible_range(x)
}
\arguments{
\item{x}{An \code{Item}.}
}
\value{
A \code{TimeRange}.
}
\description{
Visible range of an Item (including handles into transitions)
}
